projects/template/template_fastapi_app/requirements.txt
# FastAPI
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
pydantic-settings==2.0.3
email-validator==2.0.0.post2

# Database
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Google Cloud
google-cloud-pubsub==2.18.4

# OpenTelemetry
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-sqlalchemy==0.41b0
opentelemetry-instrumentation-logging==0.41b0
opentelemetry-instrumentation-requests==0.41b0
opentelemetry-instrumentation-httpx==0.41b0

# Testing
pytest==7.4.2
httpx==0.24.1

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3 